requirements.txt
# Core/Utility
numpy>=1.20.0
pandas>=1.3.0
python-dotenv>=0.19.0
tqdm
PyYAML
requests
colorama

# LLM Interaction
google-generativeai
openai

# ML/Analysis
scikit-learn>=1.0.0
datasets